% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{aa_convert}
\alias{aa_convert}
\title{Convert amino acid codes quickly}
\usage{
aa_convert(x)
}
\arguments{
\item{x}{Character vector to convert.}
}
\description{
Convert amino acid codes quickly
}
