% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_pkg.R
\docType{package}
\name{antigen.garnish}
\alias{antigen.garnish}
\alias{antigen.garnish-package}
\title{\pkg{antigen.garnish}: ensemble neoepitope prediction from DNA variants in R.}
\description{
An R package for \href{http://science.sciencemag.org/content/348/6230/69}{neoepitope} analysis that takes human or murine DNA missense mutations, insertions, deletions, and gene fusions and performs ensemble neoepitope prediction using 7 algorithms. Input is a VCF file or table of peptides. Output is ranked neoepitopes and a summary of neoepitope load and fitness by sample. Neoepitopes are ranked by MHC I/II binding affinity, clonality, RNA expression, dissimilarity to the normal peptidome, and similarity to known immunogenic antigens. \href{http://antigen-garnish-presentation.s3-website-us-east-1.amazonaws.com}{More information.}
}
\details{
Advantages
\enumerate{
\item \strong{Thoroughness}:
\itemize{
\item missense mutations, insertions, deletions, and gene fusions
\item human and mouse
\item ensemble MHC class I/II binding prediction using \href{https://github.com/hammerlab/mhcflurry}{mhcflurry}, \href{https://github.com/KarchinLab/mhcnuggets}{mhcnuggets}, \href{http://www.cbs.dtu.dk/services/NetMHC/}{netMHC}, \href{http://www.cbs.dtu.dk/services/NetMHCII/}{netMHCII}, \href{http://www.cbs.dtu.dk/services/NetMHCpan/}{netMHCpan} and \href{http://www.cbs.dtu.dk/services/NetMHCIIpan/i}{netMHCIIpan}
\item ranked by
}
\itemize{
\item MHC I/II binding affinity
\item clonality
\item RNA expression
\item dissimilarity to the normal peptidome
\item similarity to known immunogenic antigens
}
\item \strong{Speed and simplicity}:
\itemize{
\item 1000 variants are ranked in a single step in less than five minutes
}
\item \strong{Integration with R/Bioconductor}
\itemize{
\item upstream/VCF processing
\item exploratory data analysis, visualization
}
}
}
\section{Manifest}{

\itemize{
\item \code{garnish_variants}: process missense / indel VCF variants from \href{http://snpeff.sourceforge.net/}{SnpEff}
\item \code{garnish_jaffa}: process gene fusions from \href{https://github.com/Oshlack/JAFFA}{JAFFA}
\item \code{garnish_affinity}: perform ensemble neoepitope prediction
\item \code{garnish_summary}: summarize and rank results
\item \code{garnish_antigens}: list top neoepitope sequences, sources, and properties for each \href{https://github.com/lima1/PureCN}{clone} and sample
\item \code{garnish_plot}: generate summary plots
}
}

\section{Bug, issues, feedback}{

Please report bugs and issues and provide feedback via \href{https://github.com/andrewrech/antigen.garnish/issues}{Github} or to \href{mailto:rech@rech.io}{rech@rech.io}.
}

