% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{check_dep_versions}
\alias{check_dep_versions}
\title{Internal function to check for dev version of magrittr.}
\usage{
check_dep_versions()
}
\description{
Internal function to check for dev version of magrittr.
}
