% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{check_pred_tools}
\alias{check_pred_tools}
\title{Internal function to check for netMHC tools, mhcflurry and mhcnuggets}
\usage{
check_pred_tools()
}
\description{
Internal function to check for netMHC tools, mhcflurry and mhcnuggets
}
