% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{configure_netMHC_tools}
\alias{configure_netMHC_tools}
\title{Internal function to configure netMHC suite of tools}
\usage{
configure_netMHC_tools(dir)
}
\arguments{
\item{dir}{Character vector. Path to \code{antigen.garnish} data directory.}
}
\description{
Internal function to configure netMHC suite of tools
}
