% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{detect_mhc}
\alias{detect_mhc}
\title{Internal function to replace MHC with matching prediction tool MHC syntax}
\usage{
detect_mhc(x, alleles)
}
\arguments{
\item{x}{Vector of HLA types named for program to convert to.}

\item{alleles}{Data table of 2 columns, 1. formatted allele names 2. prediction tool name (e.g. mhcflurry, netMHC).}
}
\description{
Internal function to replace MHC with matching prediction tool MHC syntax
}
