% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{extract_cDNA}
\alias{extract_cDNA}
\title{Internal function to extract cDNA changes from HGVS notation}
\usage{
extract_cDNA(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Internal function to extract cDNA changes from HGVS notation
}
