% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{foreignness_score}
\alias{foreignness_score}
\title{Return foreignness_scores for a vector of peptides.}
\usage{
foreignness_score(v, db)
}
\arguments{
\item{v}{Character. Vector of nmers.}

\item{db}{Character. One of c("mouse", "human")}
}
\value{
Data table of nmers and corresponding foreignness_score values.
}
\description{
Return foreignness_scores for a vector of peptides.
}
