% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{ftrans}
\alias{ftrans}
\title{Translate cDNA to peptide fast}
\usage{
ftrans(v)
}
\arguments{
\item{v}{cDNA character vector without ambiguous bases.}
}
\description{
Translate cDNA to peptide fast
}
