% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{garnish_affinity}
\alias{garnish_affinity}
\title{Perform neoantigen prediction.}
\usage{
garnish_affinity(
  dt = NULL,
  path = NULL,
  binding_cutoff = 500,
  counts = NULL,
  min_counts = 1,
  peptide_length = 15:8,
  blast = TRUE,
  save = TRUE,
  remove_wt = TRUE
)
}
\arguments{
\item{dt}{Data table. Input data table from \code{garnish_variants}, or a data table in the correct form (see \href{https://github.com/immune-health/antigen.garnish}{Github README}).}

\item{path}{Path to input \code{csv} or \code{tsv} file.}

\item{binding_cutoff}{Numeric. Maximum consensus MHC-binding affinity that will be passed for IEDB and dissimilarity analysis. Default is 500 (nM). Note: If a peptide binds to any MHC allele in the table below this threshold, foreignness score and dissimilarity will be returned for all rows with that peptide.}

\item{counts}{Optional. A file path to a \code{csv} or \code{tsv} RNA count matrix. The first column must contain Ensembl transcript ids. All samples in the input table must be present in the count matrix.}

\item{min_counts}{Integer. The minimum number of estimated read counts for a transcript to be considered for neoantigen prediction. Default is 1.}

\item{peptide_length}{Numeric vector. Length(s) of peptides to create.}

\item{blast}{Logical. Run \code{BLASTp} to find wild-type peptide and known IEDB matches?}

\item{save}{Logical. Save a copy of garnish_affinity output to the working directory as "ag_output.txt"? Default is \code{TRUE}.}

\item{remove_wt}{Logical. Check all \code{nmer}s generated against wt peptidome and remove matches? Default is \code{TRUE}. If investigating wild-type sequences, set this to \code{FALSE}.}
}
\value{
A data table of binding predictions including:
\itemize{
\item \strong{cDNA_seq}: mutant cDNA sequence
\item \strong{cDNA_locs}: starting index of mutant cDNA
\item \strong{cDNA_locl}: ending index of mutant cDNA
\item \strong{cDNA_type}: netMHC prediction tool output
\item \strong{frameshift}: frameshift variant?
\item \strong{coding}: wt cDNA sequence
\item \strong{coding_mut}: mutant cDNA sequence
\item \strong{pep_type}: type of peptide
\item \strong{pep_mut}: mutant peptide sequence
\item \strong{pep_wt}: wt peptide sequence
\item \strong{mismatch_s}: starting index of mutant peptide sequence
\item \strong{mismatch_l}: ending index of mutant peptide sequence
\item \strong{mutant_index}: index of mutant peptide
\item \strong{nmer}: nmer for prediction
\item \strong{nmer_i}: index of nmer in sliding window
\item \strong{_net}: netMHC prediction tool output
\item \strong{mhcflurry_}: mhcflurry_ prediction tool output
\item \strong{DAI}: Differential agretopicity index of missense and corresponding wild-type peptide. Differential agretopicty is the ratio of MHC binding afinity between mutant and corresponding normal peptide, with higher values indicating greater relative binding of the mutant peptide.
\item \strong{BLAST_A}: Ratio of consensus binding affinity of mutant peptide / closest single AA mismatch from blastp results. Returned only if \code{blast = TRUE}.
}

antigen.garnish quality analysis metric results:
\itemize{
\item \strong{Ensemble_score}: average value of MHC binding affinity from all prediction tools.
\item \strong{foreignness_score}: Neoantigen foreignness threshold. Value of 0 to 1 indicating the TCR recognition probability, calculated by summing alignments in IEDB immunogenic peptides, with 1 indicating greater homology to immunogenic peptides.
\item \strong{IEDB_anno}: The best alignment from the IEDB database queried for the sample if applicable.
\item \strong{min_DAI}: Minimum of value of BLAST_A or DAI values, to provide the most conservative proteome-wide estimate of differential binding between input and wildtype matches.
\item \strong{dissimilarity}: Value of 0 to 1 indicating alignment to the self-proteome, calculated in an analogous manner to neoanigen foreignness, with 1 indicating greater dissimilarity.
}
}
\description{
Perform ensemble neoantigen prediction on a data table of missense mutations, insertions, or deletions using netMHC and mhcflurry.
}
\details{
\itemize{
\item see \code{list_mhc} for compatible MHC allele syntax, you may also use "all_human" or "all_mouse" in the MHC column to use all supported alleles
}

Parallel cores used can be set via environment variable AG_THREADS (default: all available).
}
\references{
Richman LP, Vonderheide RH, and Rech AJ. Neoantigen dissimilarity to the self-proteome predicts immunogenicity and response to immune checkpoint blockade. Cell Systems. 2019.
Duan, F., Duitama, J., Seesi, S.A., Ayres, C.M., Corcelli, S.A., Pawashe, A.P., Blanchard, T., McMahon, D., Sidney, J., Sette, A., et al. Genomic and bioinformatic profiling of mutational neoepitopes reveals new rules to predict anticancer immunogenicity. J Exp Med. 2014.

Luksza, M, Riaz, N, Makarov, V, Balachandran VP, et al. A neoepitope fitness model predicts tumour response to checkpoint blockade immunotherapy. Nature. 2017.
Rech AJ, Balli D, Mantero A, Ishwaran H, Nathanson KL, Stanger BZ, Vonderheide RH. Tumor immunity and survival as a function of alternative neopeptides in human cancer. Clinical Cancer Research, 2018.

Wells DK, van Buuren MM, Dang KK, Hubbard-Lucey VM, Sheehan KCF, Campbell KM, Lamb A, Ward JP, Sidney J, Blazquez AB, Rech AJ, Zaretsky JM, Comin-Anduix B, Ng AHC, Chour W, Yu TV, Rizvi1 H, Chen JM, Manning P, Steiner GM, Doan XC, The TESLA Consortium, Merghoub T, Guinney J, Kolom A, Selinsky C, Ribas A, Hellmann MD, Hacohen N, Sette A, Heath JR, Bhardwaj N, Ramsdell F, Schreiber RD, Schumacher TN, Kvistborg P, Defranoux N. Key Parameters of Tumor Epitope Immunogenicity Revealed Through a Consortium Approach Improve Neoantigen Prediction. Cell. 2020.
}
\seealso{
\code{\link{list_mhc}}

\code{\link{garnish_variants}}

\code{\link{garnish_antigens}}
}
