% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_antigens}
\alias{garnish_antigens}
\title{List top neoantigens for each sample and/or by clones within each sample using TESLA criteria for recognition features of immunogenic peptides.}
\usage{
garnish_antigens(dt)
}
\arguments{
\item{dt}{An output data table from \code{garnish_affinity}, either a data table object or path to a file.}
}
\value{
A data table with all neoantigens from the input table that meet at least one of the recognition features of immunogenic peptides from the \href{https://www.parkerici.org/research-project/tumor-neoantigen-selection-alliance-tesla/}{TESLA consortium}.
}
\description{
The \href{https://www.parkerici.org/research-project/tumor-neoantigen-selection-alliance-tesla/}{TESLA consortium} identified recognition features of immunogenic peptides. This function applies each of those criteria and returns any neoantigen that meets MHC binding affinity (< 34nM) and annotates in the Recognition_Features column with: agretopicity (DAI > 10), foreignness (iedb_score > 10e-16), and dissimilarity > 0.
}
\seealso{
\code{\link{garnish_affinity}}

\code{\link{garnish_summary}}
}
