% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{garnish_clonality}
\alias{garnish_clonality}
\title{Internal function to integrate clonality data into output.}
\usage{
garnish_clonality(dt)
}
\arguments{
\item{dt}{Data table. Passed internally from garnish_affinity, requires allelic_fraction or cellular_fraction columns.}
}
\value{
A data table with added clonality columns:
\itemize{
\item \strong{clone_id}: rank of the clone containing the variant (highest equals larger tumor fraction).
\item \strong{cl_proportion}: estimated mean tumor fraction containing the clone.
}
}
\description{
Integrates clonality input to help prioritize neoantigens downstream.
}
