% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{garnish_dissimilarity}
\alias{garnish_dissimilarity}
\title{Return dissimilarity (to reference proteome) values for a vector of nmers.}
\usage{
garnish_dissimilarity(v, db)
}
\arguments{
\item{v}{Character. Vector of nmers.}

\item{db}{Character. One of c("mouse", "human").}
}
\value{
Placeholder data table of nmers and corresponding dissimilarity NA values.
}
\description{
Return dissimilarity (to reference proteome) values for a vector of nmers.
}
