% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_plot}
\alias{garnish_plot}
\title{Graph \code{garnish_predictions} results.}
\usage{
garnish_plot(input)
}
\arguments{
\item{input}{Output from \code{garnish_predictions} to graph. \code{input} may be a data table object, list of data tables, or file path to a rio::import-compatible file type. If a list of data tables is provided, unique plots will be generated for each data table.}
}
\value{
NULL

As a side effect: saves graph illustrating the number of neoepitopes in each sample to the working directory. The threshold for inclusion of fusion and frameshift-derived neoepitopes is \eqn{IC_{50}} < 1000nM.
}
\description{
Plot ADN, CDN, priority, frameshift, and fusion derived \code{nmers} for class I and class II MHC by \code{sample_id}.
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # download example output
   "antigen.garnish_example_jaffa_output.txt" \%T>\%
   utils::download.file(
    "http://get.rech.io/antigen.garnish_example_jaffa_output.txt", .) \%>\%

 # create plot
   antigen.garnish::garnish_plot

}

}
\seealso{
\code{\link{garnish_predictions}}

\code{\link{garnish_summary}}
}
