% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions_worker}
\alias{garnish_predictions_worker}
\title{Internal function for parallelized \code{nmer} creation.}
\usage{
garnish_predictions_worker(dt)
}
\arguments{
\item{dt}{Data table. Input data table from \code{garnish_predictions}.}
}
\description{
Internal function for parallelized \code{nmer} creation.
}
