% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_slim}
\alias{garnish_slim}
\title{Return minimal neoepitope prediction information for all peptides.}
\usage{
garnish_slim(dt)
}
\description{
Slim table description:
}
\examples{
\dontrun{
library(magrittr)
library(data.table)
library(antigen.garnish)

 # download an example VCF
 dt <- "antigen.garnish_example.vcf" \%T>\%
 utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
 antigen.garnish::garnish_variants \%>\%

 # add test MHC types
     .[, MHC := c("HLA-A*02:01 HLA-DRB1*14:67",
                  "H-2-Kb H-2-IAd",
                 "HLA-A*01:47 HLA-DRB1*03:08")] \%>\%

 # print slim predictions
 garnish_slim \%T>\%
 print
}

}
