% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_summary}
\alias{garnish_summary}
\title{Summarize epitope prediction.}
\usage{
garnish_summary(dt)
}
\arguments{
\item{dt}{Data table. Prediction output from garnish_predictions.}
}
\description{
Calculate neoepitope summary statistics over samples.
}
