% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_targets}
\alias{garnish_targets}
\title{List the dominant neoepitope sequences for each sample and by clone if possible.}
\usage{
garnish_targets(dt)
}
\value{
A data table with the dominant neoepitope per sample, and if possible per clone, in rank order of clone frequency.
}
\description{
List the dominant neoepitope sequences for each sample and by clone if possible.
}
