% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{get_cDNA}
\alias{get_cDNA}
\title{Create cDNA from hgvs nomenclature (http://varnomen.hgvs.org/)}
\usage{
get_cDNA(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Create cDNA from hgvs nomenclature (http://varnomen.hgvs.org/)
}
