% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{get_DAI_uuid}
\alias{get_DAI_uuid}
\title{Internal function to pair peptides from missense sites by a common UUID for DAI calculations}
\usage{
get_DAI_uuid(dt)
}
\arguments{
\item{dt}{Data table of nmers.}
}
\description{
Internal function to pair peptides from missense sites by a common UUID for DAI calculations
}
