% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_metadata}
\alias{get_metadata}
\title{Internal function to add metadata by \code{ensembl_transcript_id}}
\usage{
get_metadata(dt)
}
\arguments{
\item{dt}{Data table with \code{INFO} column.}
}
\description{
Internal function to add metadata by \code{ensembl_transcript_id}
}
