% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{get_pred_commands}
\alias{get_pred_commands}
\title{Internal function to create commands for neoepitope prediction.}
\usage{
get_pred_commands(dt)

get_pred_commands(dt)
}
\arguments{
\item{dt}{Data.table of predictions to run.}

\item{dt}{Data.table of predictions to run.}
}
\description{
Internal function to create commands for neoepitope prediction.

Internal function to create commands for neoepitope prediction.
}
