% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_vcf_snpeff_dt}
\alias{get_vcf_snpeff_dt}
\title{Internal function to extract SnpEff annotation information to a data table.}
\usage{
get_vcf_snpeff_dt(dt)
}
\arguments{
\item{dt}{Data table with INFO column from a SnpEff-annotated VCF file.}
}
\description{
Internal function to extract SnpEff annotation information to a data table.
}
