% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_snpeff}
\alias{get_snpeff}
\title{Internal function to extract SnpEff annotation information to a data table.}
\usage{
get_snpeff(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Internal function to extract SnpEff annotation information to a data table.
}
