% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_vcf_info_dt}
\alias{get_vcf_info_dt}
\title{Internal function to extract a data table of variants with \code{INFO} fields in columns.}
\usage{
get_vcf_info_dt(vcf)
}
\arguments{
\item{vcf}{vcfR object to extract data from.}
}
\value{
Data table of variants with \code{INFO} fields in columns.
}
\description{
Internal function to extract a data table of variants with \code{INFO} fields in columns.
}
