% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_vcf_sample_dt}
\alias{get_vcf_sample_dt}
\title{Internal function to extract a data table from sample \code{vcf} fields.}
\usage{
get_vcf_sample_dt(vcf)
}
\arguments{
\item{vcf}{vcfR object to extract data from.}
}
\value{
Data table of variants with sample level fields in columns.
}
\description{
Internal function to extract a data table from sample \code{vcf} fields.
}
