% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{get_vcf_snpeff_dt}
\alias{get_vcf_snpeff_dt}
\title{Internal function to extract SnpEff annotation information to a data table.}
\usage{
get_vcf_snpeff_dt(dt)
}
\arguments{
\item{dt}{Data table with character vector \code{ANN} column, from \code{vcf} file.}
}
\value{
Data table with the \code{ANN} column parsed into additional rows.
}
\description{
Internal function to extract SnpEff annotation information to a data table.
}
