% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{\%exclude\%}
\alias{\%exclude\%}
\title{Convenience inflix operator to return vector elements excluding those matching a regular expression.}
\usage{
vector \%exclude\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\value{
A vector of elements in \code{vector} not matching \code{pattern}.
}
\description{
Convenience inflix operator to return vector elements excluding those matching a regular expression.
}
