% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{\%include\%}
\alias{\%include\%}
\title{Convenience inflix operator to return vector elements matching a regular expression.}
\usage{
vector \%include\% pattern
}
\arguments{
\item{vector}{Vector.}

\item{pattern}{Pattern.}
}
\value{
A vector of elements in \code{vector} matching \code{pattern}.
}
\description{
Convenience inflix operator to return vector elements matching a regular expression.
}
