% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{iedb_score}
\alias{iedb_score}
\title{Return iedb_scores for a vector of nmers.}
\usage{
iedb_score(v, db)
}
\arguments{
\item{v}{Character. Vector of nmers.}

\item{db}{Character. One of c("mouse", "human")}
}
\value{
Data table of nmers and corresponding iedb_score values.
}
\description{
Return iedb_scores for a vector of nmers.
}
