% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{make_BLAST_uuid}
\alias{make_BLAST_uuid}
\title{Internal function to categorize mutant and wild-type peptides by similarity using \code{BLAST} to calculate neoepitope amplitude.}
\usage{
make_BLAST_uuid(dti)

make_BLAST_uuid(dti)
}
\arguments{
\item{dti}{Data table of nmers.}

\item{dti}{Data table of nmers.}
}
\description{
Internal function to categorize mutant and wild-type peptides by similarity using \code{BLAST} to calculate neoepitope amplitude.

Internal function to categorize mutant and wild-type peptides by similarity using \code{BLAST} to calculate neoepitope amplitude.
}
