% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{make_cDNA}
\alias{make_cDNA}
\title{Internal function to create cDNA from HGVS notation}
\usage{
make_cDNA(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Internal function to create cDNA from HGVS notation
}
