% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{make_DAI_uuid}
\alias{make_DAI_uuid}
\title{Internal function to pair peptides from missense sites by a common UUID for DAI calculations}
\usage{
make_DAI_uuid(dt)

make_DAI_uuid(dt)
}
\arguments{
\item{dt}{Data table of nmers.}

\item{dt}{Data table of nmers.}
}
\description{
Internal function to pair peptides from missense sites by a common UUID for DAI calculations

Internal function to pair peptides from missense sites by a common UUID for DAI calculations
}
