% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_ensembl_db.R
\name{make_ensembl_db}
\alias{make_ensembl_db}
\title{Internal function to create Ensembl transcript dataabse}
\usage{
make_ensembl_db()
}
\description{
Internal function to create the antigen.garnish data directory Ensembl transcript database file \code{GRChm38_meta.RDS} from \href{ftp://ftp.ensembl.org/pub/current_fasta/homo_sapiens/cds/}{GRCh38.cds.all.fa} Ensembl files.
}
