% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{make_sw_alignment}
\alias{make_sw_alignment}
\title{Internal function to Smith-Waterman align two vectors of peptides.}
\usage{
make_sw_alignment(col1, col2, gap_open = -11, gap_extend = -1)
}
\arguments{
\item{col1}{Character. Vector of sequences to align to col2.}

\item{col2}{Character. Vector of sequences to align to col1.}

\item{gap_open}{Numeric. Cost to open gapped alignment. Default is -11.}

\item{gap_extend}{Numeric. Cost to extend a gap. Default is -1.}
}
\description{
Internal function to Smith-Waterman align two vectors of peptides.
}
