% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{merge_pred_results}
\alias{merge_pred_results}
\title{Merge input data table and prediction results from netMHC tools, mhcflurry}
\usage{
merge_pred_results(l, dt)
}
\arguments{
\item{l}{Output list from run_netMHC}

\item{dt}{Input data table.}
}
\description{
Merge input data table and prediction results from netMHC tools, mhcflurry
}
