% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{merge_predictions}
\alias{merge_predictions}
\title{Internal function to merge input data table and prediction results}
\usage{
merge_predictions(l, dt)

merge_predictions(l, dt)
}
\arguments{
\item{l}{Output list from run_netMHC}

\item{dt}{Input data table.}

\item{l}{Output list from run_netMHC}

\item{dt}{Input data table.}
}
\description{
Internal function to merge input data table and prediction results

Internal function to merge input data table and prediction results
}
