% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{run_netMHC}
\alias{run_netMHC}
\title{Internal function to run netMHC commands.}
\usage{
run_netMHC(dt)
}
\arguments{
\item{dt}{Data table of prediction commands to run.}
}
\description{
Internal function to run netMHC commands.
}
