% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{translate_cDNA}
\alias{translate_cDNA}
\title{Internal function to translate cDNA to peptides}
\usage{
translate_cDNA(v)
}
\arguments{
\item{v}{cDNA character vector without ambiguous bases.}
}
\description{
Internal function to translate cDNA to peptides
}
