% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{write_netmhc_nmers}
\alias{write_netmhc_nmers}
\title{Internal function to output nmers for netMHC prediction to disk}
\usage{
write_netmhc_nmers(dt, type)

write_netmhc_nmers(dt, type)
}
\arguments{
\item{dt}{Data table of nmers.}

\item{type}{Character vector. Name of program to format for.}

\item{dt}{Data table of nmers.}

\item{type}{Character vector. Name of program to format for.}
}
\description{
Internal function to output nmers for netMHC prediction to disk

Internal function to output nmers for netMHC prediction to disk
}
