% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{write_nmers}
\alias{write_nmers}
\title{Write MHC prediction nmers to disk}
\usage{
write_nmers(dt, type)
}
\arguments{
\item{dt}{Data table of nmers.}

\item{type}{Character vector. Name of program to format for.}
}
\description{
Write MHC prediction nmers to disk
}
