% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_pkg.R
\docType{package}
\name{antigen.garnish}
\alias{antigen.garnish}
\title{\pkg{antigen.garnish}: ensemble neoantigen prediction from DNA variants in R.}
\description{
Github: https://github.com/immune-health/antigen.garnish
Documentation: https://neoantigens.rech.io/
}
\seealso{
\code{\link{garnish_variants}}

\code{\link{garnish_affinity}}

\code{\link{garnish_antigens}}
}
