% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_pkg.R
\docType{package}
\name{antigen.garnish}
\alias{antigen.garnish}
\title{\pkg{antigen.garnish}: tumor neoantigen prediction}
\description{
\href{https://github.com/immune-health/antigen.garnish}{Github}
\href{https://neoantigens.rech.io/}{Documentation}
}
\seealso{
\code{\link{garnish_variants}}

\code{\link{garnish_affinity}}

\code{\link{garnish_antigens}}
}
