% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_pkg.R
\docType{package}
\name{antigen.garnish}
\alias{antigen.garnish}
\alias{antigen.garnish-package}
\title{\pkg{antigen.garnish}: ensemble neoepitope prediction from DNA variants in R.}
\description{
\href{http://neoepitopes.io}{antigen.garnish} is an R package for \href{http://science.sciencemag.org/content/348/6230/69}{neoepitope} analysis that takes human or murine DNA missense mutations, insertions, deletions, and gene fusions and performs neoepitope prediction. Input is a VCF file or table of peptides. Output is neoepitopes and a summary of neoepitope load and fitness by sample.
}
\details{
Advantages
\enumerate{
\item \strong{Thoroughness}:
\itemize{
\item missense mutations, insertions, deletions, and gene fusions
\item human and mouse
\item ensemble MHC class I/II binding prediction using \href{https://github.com/hammerlab/mhcflurry}{mhcflurry}, \href{https://github.com/KarchinLab/mhcnuggets}{mhcnuggets}, \href{http://www.cbs.dtu.dk/services/NetMHC/}{netMHC}, \href{http://www.cbs.dtu.dk/services/NetMHCII/}{netMHCII}, \href{http://www.cbs.dtu.dk/services/NetMHCpan/}{netMHCpan} and \href{http://www.cbs.dtu.dk/services/NetMHCIIpan/i}{netMHCIIpan}
\item rank by dissimilarity to the normal peptidome / similarity to known immunogenic antigens
}
\item \strong{Speed and simplicity}:
\itemize{
\item 1000 variants are ranked in a single step in less than five minutes
}
\item \strong{Integration with R/Bioconductor}
\itemize{
\item upstream/VCF processing
\item exploratory data analysis, visualization
}
}
}
\section{Manifest}{

\itemize{
\item \code{garnish_variants}: process missense / indel VCF variants from \href{http://snpeff.sourceforge.net/}{SnpEff}
\item \code{garnish_jaffa}: process gene fusions from \href{https://github.com/Oshlack/JAFFA}{JAFFA}
\item \code{garnish_predictions}: perform ensemble neoepitope prediction
\item \code{garnish_summary}: summarize and rank results
\item \code{garnish_plot}: generate summary plots
\item \code{list_mhc}: list all supported MHC allele syntax
}
}

\section{Bug, issues, feedback}{

Please report bugs and issues and provide feedback via \href{https://github.com/andrewrech/antigen.garnish/issues}{Github} or to \href{mailto:andrewrech@gmail.com}{andrewrech@gmail.com}.
}

