% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_pkg.R
\docType{package}
\name{antigen.garnish}
\alias{antigen.garnish}
\alias{antigen.garnish-package}
\title{\pkg{antigen.garnish}: ensemble neoepitope prediction from DNA variants in R.}
\description{
\href{http://neoepitopes.io}{antigen.garnish} is an R package for \href{http://science.sciencemag.org/content/348/6230/69}{neoepitope} analysis that takes human or murine DNA point mutations, insertions, and deletions in VCF format and performs neoepitope prediction. Output is individual peptides and a summary by sample.
}
\details{
Advantages
\enumerate{
\item \strong{Simplicity}: summarized neoepitopes for each sample
\item \strong{Thoroughness}:
\itemize{
\item missense mutations and frameshifts
\item ensemble MHC class I/II binding prediction using \href{https://github.com/hammerlab/mhcflurry}{mhcflurry}, \href{http://www.cbs.dtu.dk/services/NetMHC/}{netMHC}, \href{http://www.cbs.dtu.dk/services/NetMHCII/}{netMHCII}, \href{http://www.cbs.dtu.dk/services/NetMHCpan/}{netMHCpan} and \href{http://www.cbs.dtu.dk/services/NetMHCIIpan/i}{netMHCIIpan}.
}
\item \strong{Speed}:
\itemize{
\item produce all possible 8-15-mer peptides from 10,000 variants in under 1 minute on a normal laptop
\item on an Amazon Web Services \code{m4.16xlarge} EC2 instance, 20,000 consensus predictions using 100+ MHC types in under 5 minutes
}
}
}
\section{Manifest}{

\itemize{
\item \code{garnish_variants}: process variants from \href{http://snpeff.sourceforge.net/}{SnpEff}
\item \code{garnish_predictions}: perform ensemble neoepitope prediction
\item \code{garnish_summary}: summarize neoepitope prediction
}
}

