% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R, R/garnish_plots.R
\name{cat_tables}
\alias{cat_tables}
\alias{cat_tables}
\title{Internal function to combine garnish_predictions outputs from garnish_jaffa
and garnish_variants or direct input to garnish_predictions.}
\usage{
cat_tables(dt1, dt2)

cat_tables(dt1, dt2)
}
\description{
Internal function to combine garnish_predictions outputs from garnish_jaffa
and garnish_variants or direct input to garnish_predictions.

Internal function to combine garnish_predictions outputs from garnish_jaffa
and garnish_variants or direct input to garnish_predictions.
}
