% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_run.R
\name{check_pred_tools}
\alias{check_pred_tools}
\title{Internal function to check for netMHC tools and mhcflurry}
\usage{
check_pred_tools(
  ag_dirs = c(paste0(getwd(), "/", "antigen.garnish"), paste0(Sys.getenv("HOME"), "/",
    "antigen.garnish"))
)
}
\arguments{
\item{ag_dirs}{Character vector. Directories to check for \code{antigen.garnish} data directory.}
}
\description{
Internal function to check for netMHC tools and mhcflurry
}
