% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{collate_netMHC}
\alias{collate_netMHC}
\title{Collate results from netMHC prediction}
\usage{
collate_netMHC(esl)
}
\arguments{
\item{esl}{List of outputs from netMHC.}
}
\description{
Collate results from netMHC prediction
}
