% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{collate_netMHC}
\alias{collate_netMHC}
\title{Internal function to collate results from netMHC prediction}
\usage{
collate_netMHC(esl)

collate_netMHC(esl)
}
\arguments{
\item{esl}{List of outputs from netMHC.}

\item{esl}{List of outputs from netMHC.}
}
\description{
Internal function to collate results from netMHC prediction

Internal function to collate results from netMHC prediction
}
