% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{detect_hla}
\alias{detect_hla}
\title{Replace HLA with matching type in netMHC format}
\usage{
detect_hla(x)
}
\arguments{
\item{x}{Vector of HLA types named for program to convert to.}
}
\description{
Replace HLA with matching type in netMHC format
}
