% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_assemble.R
\name{detect_mhc}
\alias{detect_mhc}
\title{Replace MHC with matching type in prediction tool format}
\usage{
detect_mhc(x, alleles)
}
\arguments{
\item{x}{Vector of HLA types named for program to convert to.}

\item{alleles}{Data table of 2 columns, 1. formatted allele names 2. prediction tool name (e.g. mhcflurry, mhcnuggets, netMHC).}
}
\description{
Replace MHC with matching type in prediction tool format
}
