% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{extract_cDNA}
\alias{extract_cDNA}
\title{Extract cDNA changes from hgvs nomenclature (http://varnomen.hgvs.org/)}
\usage{
extract_cDNA(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Extract cDNA changes from hgvs nomenclature (http://varnomen.hgvs.org/)
}
