% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish.antigens}
\alias{garnish.antigens}
\title{List the top neoepitope sequences for each sample and by clone if possible.}
\usage{
garnish.antigens(dt)
}
\value{
A data table with the top neoepitope per sample, and if possible per clone, in rank order of clone frequency.
}
\description{
List the top neoepitope sequences for each sample and by clone if possible.
}
