% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_antigens}
\alias{garnish_antigens}
\title{List top neoepitopes for each sample and/or by clones within each sample..}
\usage{
garnish_antigens(dt)
}
\value{
A data table with the top neoepitope per sample, and if possible per clone, in rank order of clone frequency.
}
\description{
List top neoepitopes for each sample and/or by clones within each sample..
}
