% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_jaffa.R
\name{garnish_jaffa}
\alias{garnish_jaffa}
\title{Process gene fusions and return a data table for neoepitope prediction.}
\usage{
garnish_jaffa(path, db, fasta_path)
}
\arguments{
\item{path}{Path to \code{jaffa_results.csv}.}

\item{db}{Character vector. One of \code{GRCm37}, \code{GRCm38}, \code{GRCh37}, or \code{GRCh38}.}

\item{fasta_path}{Path to \code{jaffa_results.fasta}.}
}
\value{
A data table of mutant peptides, including:
\itemize{
\item \strong{sample_id}: sample id
\item \strong{pep_mut}: mutant fusion peptide
\item \strong{mutant_index}: index of mutant peptide
\item \strong{fusion gene}: colon-separated gene names for the fusion product
\item \strong{chrom1}: chromosome of first fusion gene
\item \strong{base1}: base number of first fusion gene breakpoint
\item \strong{chrom2}: chromosome of second fusion gene
\item \strong{base2}: base number of second fusion gene breakpoint
\item \strong{fusion_uuid}: unique identifier
\item \strong{pep_wt}: wt cDNA sequence of peptide from 5' fusion gene.
\item \strong{fus_tx}: cDNA sequence of predicted fusion product
}
}
\description{
Process \href{https://github.com/Oshlack/JAFFA}{JAFFA} gene fusion \code{fasta} and \code{results.csv} output for neoepitope prediction using \code{garnish_predictions}.
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # load example jaffa output
   path <- "antigen.garnish_jaffa_results.csv" \%T>\%
     utils::download.file("http://get.rech.io/antigen.garnish_jaffa_results.csv", .)
   fasta_path <- "antigen.garnish_jaffa_results.fasta" \%T>\%
     utils::download.file("http://get.rech.io/antigen.garnish_jaffa_results.fasta", .)

 # get predictions
   dt <- antigen.garnish::garnish_jaffa(path, db = "GRCm38", fasta_path) \%>\%

 # add MHC info with antigen.garnish::list_mhc() compatible names
   .[, MHC := "H-2-Kb"] \%>\%

 # get predictions
   antigen.garnish::garnish_predictions \%>\%

 # summarize predictions
   antigen.garnish::garnish_summary \%T>\%
   print
}

}
\seealso{
\code{\link{garnish_predictions}}
}
