% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_plot}
\alias{garnish_plot}
\title{Graph \code{garnish_affinity} results.}
\usage{
garnish_plot(input, ext = "pdf")
}
\arguments{
\item{input}{Output from \code{garnish_affinity} to graph. \code{input} may be a data table object, list of data tables, or file path to a rio::import-compatible file type. If a list of data tables is provided, unique plots will be generated for each data table.}

\item{ext}{File extension passed to ggplot2::ggsave, default is "pdf".}
}
\value{
NULL

As a side effect: saves graph illustrating the number of neoantigens in each sample to the working directory. The threshold for inclusion of neoantigens is \eqn{IC_{50}} < 500nM.
}
\description{
Plot ADN, CDN, priority, frameshift, and fusion derived \code{nmers} for class I and class II MHC by \code{sample_id}.
}
\examples{
\dontrun{

# load example output
 dir <- system.file(package = "antigen.garnish") \%>\%
   file.path(., "extdata/testdata")

   "antigen.garnish_PureCN_example_output.txt" \%>\%
   file.path(dir, .) \%>\%

# create plot
   garnish_plot

}

}
\seealso{
\code{\link{garnish_affinity}}

\code{\link{garnish_summary}}
}
