% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R, R/garnish_plots.R
\name{garnish_plot}
\alias{garnish_plot}
\alias{garnish_plot}
\title{Graph summary results using ggplot2.}
\usage{
garnish_plot(dt = NULL, jdt = NULL, save = TRUE)

garnish_plot(dt = NULL, jdt = NULL, save = TRUE)
}
\arguments{
\item{dt}{Data table. Prediction output from \code{garnish_predictions}. Either dt or jdt must be provided as input.}

\item{jdt}{Data table. Prediction output from \code{garnish_jaffa} %>% \code{garnish_predictions}. Either dt or jdt must be provided as input.}

\item{save}{Optional: Should the plot be saved to the working directory as antigen.garnish_summary.pdf? Default is TRUE.}

\item{dt}{Data table. Prediction output from \code{garnish_predictions}. Either dt or jdt must be provided as input.}

\item{jdt}{Data table. Prediction output from \code{garnish_jaffa`` %>%}garnish_predictions`. Either dt or jdt must be provided as input.}

\item{save}{Optional: Should the plot be saved to the working directory as antigen.garnish_summary.pdf? Default is TRUE.}
}
\value{
A list of ggplot2 objects and, if save = TRUE, saves each of these objects pdfs in the working directory.
The first plot is a column graph indicating the number of peptides in each sample classified as ADN, CDN, priority, frameshift-derived, and fusion-derived neoepitopes
(if any) for each sample, faceted by Class I vs Class II MHC. The threshold for fusion and frameshift derived neoepitopes is Consensus_scores < 1000nM.
See `garnish_summary`` documentation for further explanation of neoepitope classification of ADN, CDN, and priority.

The second plot and third plots are returned only if frameshift and/or fusion mutants (respectively) are present in the input table. These plot
the number of peptides per sample by MHC class I or II and binned by binding affinity (1000 - 500nM, 500 - 50nM, and < 50 nM).

A list of ggplot2 objects and, if save = TRUE, saves each of these objects pdfs in the working directory.
The first plot is a column graph indicating the number of peptides in each sample classified as ADN, CDN, priority, frameshift-derived, and fusion-derived neoepitopes
(if any) for each sample, faceted by Class I vs Class II MHC. The threshold for fusion and frameshift derived neoepitopes is Consensus_scores < 1000nM.
See `garnish_summary`` documentation for further explanation of neoepitope classification of ADN, CDN, and priority.

The second plot and third plots are returned only if frameshift and/or fusion mutants (respectively) are present in the input table. These plot
the number of peptides per sample by MHC class I or II and binned by binding affinity (1000 - 500nM, 500 - 50nM, and < 50 nM).
}
\description{
Plots ADN, CDN, priority, frameshift, and fusion derived neoepitopes for Class I and Class II MHC by sample.
See \code{garnish_summary} for more information on neoepitope classification by antigen.garnish.

Plots ADN, CDN, priority, frameshift, and fusion derived neoepitopes for Class I and Class II MHC by sample.
See \code{garnish_summary} for more information on neoepitope classification by antigen.garnish.
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # download an example VCF
   g <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   antigen.garnish::garnish_variants \%>\%

 # add test MHC types
     .[, MHC := c("HLA-A*02:01 HLA-DRB1*14:67",
                  "H-2-Kb H-2-IAd",
                 "HLA-A*01:47 HLA-DRB1*03:07")] \%>\%

 # predict neoepitopes
   antigen.garnish::garnish_predictions \%>\%

 # plot it
   antigen.garnish::garnish_plot
  
 # here is our ggplot object that can be further customized 
  class(g)
 
 # save the plot with a custom ggplot theme 
 g + my_ggplot_custom_theme_object \%>\% cowplot::ggsave("antigen.garnish_custom_themed_plot.pdf")
  
 # show location of where default plot was saved
  list.files(pattern = "summary\\\\.pdf", full.names = TRUE)   
  
}

\dontrun{
library(magrittr)
library(antigen.garnish)

 # download an example VCF
   g <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   antigen.garnish::garnish_variants \%>\%

 # add test MHC types
     .[, MHC := c("HLA-A*02:01 HLA-DRB1*14:67",
                  "H-2-Kb H-2-IAd",
                 "HLA-A*01:47 HLA-DRB1*03:07")] \%>\%

 # predict neoepitopes
   antigen.garnish::garnish_predictions \%>\%

 # plot it
   antigen.garnish::garnish_plot
  
 # here is our ggplot object that can be further customized 
  class(g)
 
 # save the plot with a custom ggplot theme 
 g + my_ggplot_custom_theme_object \%>\% cowplot::ggsave("antigen.garnish_custom_themed_plot.pdf")
  
 # show location of where default plot was saved
  list.files(pattern = "summary\\\\.pdf", full.names = TRUE)   
  
}

}
