% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions}
\alias{garnish_predictions}
\title{Performs epitope prediction.}
\usage{
garnish_predictions(dt, assemble = TRUE, generate = TRUE, predict = TRUE)
}
\arguments{
\item{dt}{Data frame. Input data frame from garnish_variants. \href{http://get.rech.io/antigen.garnish_example_input.txt}{Example.}}

\item{assemble}{Logical. Assemble data table?}

\item{generate}{Logical. Generate peptides?}

\item{predict}{Logical. Predict binding affinities?}
}
\value{
A data table of neoepitopes.
}
\description{
Performs epitope prediction on a data table of missense mutations.
}
\examples{
library(magrittr)

 # download an example VCF
   dt <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   antigen.garnish::garnish_variants \%>\%

 # predict neoepitopes
   antigen.garnish::garnish_predictions \%T>\%
   str

}
