% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions}
\alias{garnish_predictions}
\title{Performs epitope prediction.}
\usage{
garnish_predictions(dt, assemble = TRUE, generate = TRUE, predict = TRUE,
  humandb = "GRCh38", mousedb = "GRCm38")
}
\arguments{
\item{dt}{Data table. Input data table from garnish_variants. \href{http://get.rech.io/antigen.garnish_example_input.txt}{Example}. Required columns: sample_id, nensembl_transcript_id (e.g. "ENST00000311936"), cDNA_change HGVS notation (e.g. "c.718T>A"), space-separated MHC (e.g. "HLA-A*02:01 H-2-Kb HLA-DRB1*03:01").}

\item{assemble}{Logical. Assemble data table?}

\item{generate}{Logical. Generate peptides?}

\item{predict}{Logical. Predict binding affinities?}

\item{humandb}{Character vector. One of "GRCh37" or "GRCh38".}

\item{mousedb}{Character vector. One of "GRCm37" or "GRCm38".}
}
\value{
A data table of neoepitopes.
}
\description{
Performs epitope prediction on a data table of missense mutations.
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # download an example VCF
   dt <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   garnish_variants \%>\%

 # predict neoepitopes
   garnish_predictions \%T>\%
   str
}

}
