% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions}
\alias{garnish_predictions}
\title{Performs epitope prediction.}
\usage{
garnish_predictions(dt, assemble = TRUE, generate = TRUE, predict = TRUE)
}
\arguments{
\item{dt}{Data frame. Input data frame from garnish_variants. Must contain four columns: sample_id, transcript_affected (ensembl_transcript_id), aa_mutation (single letter abbreviations), and space-seperated MHC (MHC type, e.g. 'HLA-A*02:01 HLA-A*03:01', 'H-2-Kb H-2-Kb' or HLA-DRB1*11:07 [second type]')}

\item{assemble}{Logical. Assemble data table?}

\item{generate}{Logical. Generate peptides and commands?}

\item{predict}{Logical. Predict binding affinities?}
}
\value{
A data table of neoepitopes.
}
\description{
Performs epitope prediction on a data table of missence mutations.
}
