% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{garnish_predictions}
\alias{garnish_predictions}
\title{Performs epitope prediction.}
\usage{
garnish_predictions(dt, assemble = TRUE, generate = TRUE, predict = TRUE,
  humandb = "GRCh38", mousedb = "GRCm38")
}
\arguments{
\item{dt}{Data table. Input data table from garnish_variants. Required columns: sample_id, ensembl_transcript_id (e.g. \code{ENST00000311936}), cDNA_change \href{http://varnomen.hgvs.org/recommendations/DNA/}{HGVS} notation (e.g. \code{c.718T>A}), space-separated MHC (e.g. \code{HLA-A*02:01 H-2-Kb HLA-DRB1*03:01}).}

\item{assemble}{Logical. Assemble data table?}

\item{generate}{Logical. Generate peptides?}

\item{predict}{Logical. Predict binding affinities?}

\item{humandb}{Character vector. One of "GRCh37" or "GRCh38".}

\item{mousedb}{Character vector. One of "GRCm37" or "GRCm38".}
}
\value{
A data table of binding predictions including:
\itemize{
\item \strong{cDNA_seq}: mutant cDNA sequence
\item \strong{cDNA_locs}: starting index of mutant cDNA
\item \strong{cDNA_locl}: ending index of mutant cDNA
\item \strong{cDNA_type}: netMHC prediction tool output
\item \strong{frameshift}: frameshift variant?
\item \strong{coding}: wt cDNA sequence
\item \strong{coding_mut}: mutant cDNA sequence
\item \strong{pep_type}: type of peptide for this row
\item \strong{pep_mut}: mutant cDNA sequence
\item \strong{pep_wt}: wt cDNA sequence
\item \strong{mismatch_s}: starting index of mutant peptide sequence
\item \strong{mismatch_l}: ending index of mutant peptide sequence
\item \strong{mutant_loc}: index of mutant peptide for this row
\item \strong{nmer}: nmer for prediction
\item \strong{nmer_i}: index of nmer in sliding window
\item \strong{*_net}: netMHC prediction tool output
\item \strong{mhcflurry_*}: netMHC prediction tool output
}

as well as a transcript description:
\itemize{
\item description
\item start_position
\item end_position
\item transcript_end
\item transcript_length
\item transcript_start
\item peptide
\item refseq_mrna
}
}
\description{
Performs epitope prediction on a data table of missense mutations.
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # download an example VCF
   dt <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   garnish_variants \%>\%

 # predict neoepitopes
   garnish_predictions \%T>\%
   str
}

\dontrun{
# using an existing data table

library(data.table)
library(magrittr)
library(antigen.garnish)

 dt <- data.table::data.table(
          sample_id = "test",
          ensembl_transcript_id =
          c("ENSMUST00000128119",
            "ENSMUST00000044250",
            "ENSMUST00000018743"),
          cDNA_change = c("c.4988C>T",
                          "c.1114T>G",
                          "c.718T>A"),
          MHC = c("HLA-A*02:01 HLA-DRB1*14:67",
                  "H-2-Kb H-2-IAd",
                  "HLA-A*01:47 HLA-DRB1*03:08")) \%>\%
 garnish_predictions \%T>\%
 str
}
}
