% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions_worker}
\alias{garnish_predictions_worker}
\title{Parallelized worker function for garnish_predictions}
\usage{
garnish_predictions_worker(dt)
}
\arguments{
\item{dt}{Data table. Input data table from garnish_predictions.}
}
\value{
A data table summary of neoepitope by sample_id.
}
\description{
Parallelized worker function for garnish_predictions
}
