% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_predictions_worker}
\alias{garnish_predictions_worker}
\title{Parallelized worker function for garnish_predictions}
\usage{
garnish_predictions_worker(basepep_dt)
}
\arguments{
\item{basepep_dt}{Data table. Input data frame from garnish_predictions.}
}
\description{
Parallelized worker function for garnish_predictions
}
