% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_summary}
\alias{garnish_summary}
\title{Summarize epitope prediction.}
\usage{
garnish_summary(dt)
}
\arguments{
\item{dt}{Data table. Prediction output from garnish_predictions.}
}
\description{
Calculate neoepitope summary statistics over samples.
}
\examples{
library(magrittr)

 # download an example VCF
   dt <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   antigen.garnish::garnish_variants \%>\%

 # predict neoepitopes
   antigen.garnish::garnish_predictions \%>\%

 # summarize predictions
   antigen.garnish::garnish_summary \%T>\%
   print

 # example output
   dt <- data.table::fread(
    "http://get.rech.io/antigen.garnish_example_summary.txt") \%T>\%
    str

}
