% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_targets}
\alias{garnish_targets}
\title{List the dominant neoepitope sequences for each clone in each sample.}
\usage{
garnish_targets(dt)
}
\value{
A data table with the dominant neoepitope per clone per sample, in rank order of clone frequency.
}
\description{
Requires that garnish_score was computed from allelic or clonal proportion data.
}
