% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_variants}
\alias{garnish_variants}
\title{Intakes variants from MuTect2/Strelka to SnpEff pipeline and returns a standard data frame for epitope prediction.}
\usage{
garnish_variants(vcfs)
}
\arguments{
\item{vcfs}{Character vector. VFC files to import.}
}
\description{
Intakes filtered missense mutations with known Ensembl transcript identifier from a MuTect2/Strelka to SnpEff variant annotation pipeline and returns a standard data frame of intersects.
}
\examples{

}
