% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{garnish_variants}
\alias{garnish_variants}
\title{Intakes variants and returns an intersected data table for epitope prediction.}
\usage{
garnish_variants(vcfs)
}
\arguments{
\item{vcfs}{Character vector. VFC files to import.}
}
\value{
A data table of intersected protein coding variants for neoepitope prediction.
}
\description{
Process raw variants annotated with \href{https://github.com/pcingola/SnpEff}{SnpEff}. VCFs from matched samples are optionally intersected. \href{https://github.com/broadinstitute/gatk}{MuTect2}/\href{https://github.com/Illumina/strelka}{Strelka2}-derived VCFs are filtered for high confidence variants prior to intersection.
}
