% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_io.R
\name{garnish_variants}
\alias{garnish_variants}
\title{Process VCF variants and return a data table for epitope prediction.}
\usage{
garnish_variants(vcfs, intersect = TRUE)
}
\arguments{
\item{vcfs}{Paths to VFC files to import.}

\item{intersect}{Logical. Return only the intersection of variants in multiple \code{vcfs} with identical sample names? Intersection performed on \code{SnpEff} annotations. One \code{vcf} file per somatic variant caller-input samples pair is required.}
}
\value{
A data table with one unique SnpEff variant annotation per row, including:
\itemize{
\item \strong{sample_id}: sample identifier constructed from input \code{.bam} file names
\item \strong{se}: SnpEff annotation
\item \strong{effect_type}: SnpEff effect type
\item \strong{ensembl_transcript_id}: transcript effected
\item \strong{ensembl_gene_id}: gene effected
\item \strong{protein_change}: protein change in \href{http://varnomen.hgvs.org/recommendations/DNA/}{HGVS} format
\item \strong{cDNA_change}: cDNA change in \href{http://varnomen.hgvs.org/recommendations/protein/}{HGVS} format
\item \strong{protein_coding}: is the variant protein coding?
}
}
\description{
Process paired tumor-normal VCF variants annotated with \href{http://snpeff.sourceforge.net/}{SnpEff} for neoepitope prediction using \code{garnish_predictions}. VCFs from matched samples can be optionally intersected to select only variants present across input files.
}
\details{
Recommended somatic variant callers: \href{https://github.com/broadinstitute/gatk}{MuTect2}, \href{https://github.com/Illumina/strelka}{Strelka2}
}
\examples{
\dontrun{
library(magrittr)
library(antigen.garnish)

 # download an example VCF
   dt <- "antigen.garnish_example.vcf" \%T>\%
   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%

 # extract variants
   garnish_variants \%T>\%
   str
}

}
\references{
Callari M, Sammut SJ, De Mattos-Arruda L, Bruna A, Rueda OM, Chin SF, and Caldas C. 2017. Intersect-then-combine approach: improving the performance of somatic variant calling in whole exome sequencing data using multiple aligners and callers. Genome Medicine. 9:35.
}
\seealso{
\code{\link{garnish_predictions}}
}
