% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{get_metadata}
\alias{get_metadata}
\title{Add metadata using ensembl_transcript_ID}
\usage{
get_metadata(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Add metadata using ensembl_transcript_ID
}
