% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{get_pred_commands}
\alias{get_pred_commands}
\title{Collate results from netMHC prediction}
\usage{
get_pred_commands(dt)
}
\arguments{
\item{dt}{Data.table of predictions to run.}
}
\description{
Collate results from netMHC prediction
}
