% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{get_pred_commands}
\alias{get_pred_commands}
\title{Internal function to create commands for neoantigen prediction.}
\usage{
get_pred_commands(dt)
}
\arguments{
\item{dt}{Data.table of predictions to run.}
}
\description{
Internal function to create commands for neoantigen prediction.
}
