% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{get_snpeff_annot}
\alias{get_snpeff_annot}
\title{Extract snpeff annotation information to a data table.}
\usage{
get_snpeff_annot(dt)
}
\arguments{
\item{dt}{Data table with INFO column.}
}
\description{
Extract snpeff annotation information to a data table.
}
