% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R,
%   R/antigen.garnish_predict.sync-conflict-20171216-110648-X345HBM.R
\name{get_ss_str}
\alias{get_ss_str}
\title{Parallelized function to create a space-separated character string between two integers}
\usage{
get_ss_str(x, y)

get_ss_str(x, y)
}
\arguments{
\item{x}{Integer. Starting integer.}

\item{y}{Integer. Ending integer.}

\item{x}{Integer. Starting integer.}

\item{y}{Integer. Ending integer.}
}
\description{
Parallelized function to create a space-separated character string between two integers

Parallelized function to create a space-separated character string between two integers
}
