% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{make_BLAST_uuid}
\alias{make_BLAST_uuid}
\title{Internal function to pair peptides with similar wild-type peptides using BLAST to calculate neoepitope amplitude.}
\usage{
make_BLAST_uuid(dti)
}
\arguments{
\item{dti}{Data table of nmers.}
}
\description{
Internal function to pair peptides with similar wild-type peptides using BLAST to calculate neoepitope amplitude.
}
