% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{make_nmers}
\alias{make_nmers}
\title{Internal function for parallelized \code{nmer} creation.}
\usage{
make_nmers(dt, plen = 8:15)
}
\arguments{
\item{dt}{Data table. Input data table from \code{garnish_affinity}.}

\item{plen}{Numeric vector. Length(s) of peptides to create.}
}
\description{
Internal function for parallelized \code{nmer} creation.
}
