% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish.R
\name{merge_predictions}
\alias{merge_predictions}
\title{Internal function to merge input data table and prediction results from netMHC tools, mhcflurry}
\usage{
merge_predictions(l, dt)
}
\arguments{
\item{l}{Output list from run_netMHC}

\item{dt}{Input data table.}
}
\description{
Internal function to merge input data table and prediction results from netMHC tools, mhcflurry
}
