% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{write_mhcnuggets_nmers}
\alias{write_mhcnuggets_nmers}
\title{Internal function to output nmers for mhcnuggets prediction to disk}
\usage{
write_mhcnuggets_nmers(dt, alleles)
}
\arguments{
\item{dt}{Data table of nmers.}

\item{alleles}{Data table of 2 columns, 1. alleles properly formatted mhcnuggets.}
}
\description{
Internal function to output nmers for mhcnuggets prediction to disk
}
