% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigen.garnish_predict.R
\name{write_netmhc_nmers}
\alias{write_netmhc_nmers}
\title{Internal function to output nmers for MHC prediction to disk}
\usage{
write_netmhc_nmers(dt, type)
}
\arguments{
\item{dt}{Data table of nmers.}

\item{type}{Character vector. Name of program to format for.}
}
\description{
Internal function to output nmers for MHC prediction to disk
}
